+++ auth0/src/main/java/com/auth0/android/management/UsersAPIClient.java
@@ -133,8 +133,9 @@ public class UsersAPIClient {
      *
      * @param context a valid Context
      * @param token   of the primary identity
+     * @deprecated This method will be removed in the next version. Please use {@link UsersAPIClient(Auth0, String)}.
      */
-    //TODO: Remove this constructor. Replacement: UsersAPIClient(Auth0(context), "token")
+    @Deprecated
     public UsersAPIClient(@NonNull Context context, @NonNull String token) {
         this(new Auth0(context), token);
     }