+++ auth0/src/test/java/com/auth0/android/provider/WebAuthProviderTest.java
@@ -1352,6 +1352,7 @@ public class WebAuthProviderTest {
         mockAPI.shutdown();
     }
 
+    @SuppressWarnings("deprecation")
     @Test
     public void shouldResumeLoginIgnoringEmptyCustomIDTokenVerificationIssuer() throws Exception {
         PKCE pkce = Mockito.mock(PKCE.class);